Add files via upload
bulk_file.zip is all of the graphics that I made for my website. These are not all wallpapers, some were used for buttons, some were uploaded or made for the website but were changed at a later date and some were made but never really used. So feel free to download it and play around with them if you want!